requirements.txt
hikari-lightbulb~=2.2.0
python-dotenv~=0.19.2
imgurpython~=1.1.7
psutil~=5.9.0
hurry.filesize~=0.9
uptime~=3.0.1